requirements.txt
black
flake8
isort
pip
python-dotenv
pytesseract
tabula-py==2.3.0
#spacy==3.2.1
nltk==3.6.7
pdfplumber>=0.5.28
pymupdf==1.24.9
pypdf==3.9.1
pdfrw==0.4
streamlit
transformers
tensorflow
camelot-py[cv]
opencv-python-headless
timm
psutil
torch
